requirements.txt
absl-py==0.1.9
alembic==0.9.8
asn1crypto==0.22.0
attrs==17.3.0
Augmentor==0.1.8
bleach==2.1.2
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
enum34==1.1.6
environment-kernels==1.1
Flask==0.12.2
future==0.16.0
futures==3.1.1
h5py==2.7.1
html5lib==1.0.1
hyperas==0.4
hyperopt==0.1
idna==2.6
imbalanced-learn==0.3.2
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.4.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.1.0
jupyter-latex-envs==1.4.1
jupyter-nbextensions-configurator==0.4.0
jupyterhub==0.8.1
jupyterhub-systemdspawner==0.9.10
Keras==2.1.3
lxml==4.1.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nose==1.3.7
notebook==5.4.0
numpy==1.14.0
olefile==0.44
opencv-python==3.4.0.12
pamela==0.3.0
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.6.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
py==1.5.2
pycparser==2.18
Pygments==2.2.0
pymongo==3.6.0
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.3.1
python-dateutil==2.6.1
python-editor==1.0.3
python-oauth2==1.1.0
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
requests==2.18.4
ruamel-yaml==0.11.14
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.3
tensorflow==1.1.0
tensorflow-gpu==1.1.0
tensorflow-tensorboard==1.5.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
tqdm==4.19.4
traitlets==4.3.2
tuner==0.0.2a1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.8
yapf==0.19.0
